+++ shared/versions.ts
@@ -1,7 +1,7 @@
-export const APP_VERSION = "2.3.278";
+export const APP_VERSION = "2.3.279";
 export const ANDROID_APP_VERSION = "2.3.97";
-export const ANDROID_APK_RELEASE_VERSION = "2.3.278";
-export const AGENT_VERSION = "2.2.192";
+export const ANDROID_APK_RELEASE_VERSION = "2.3.279";
+export const AGENT_VERSION = "2.3.0";
 
 export const PANEL_AGENT_COMPATIBILITY_LIMIT = 5;
 export const PANEL_AGENT_COMPATIBILITY = [